Recreate instrument history list based on sitechan
Remove dbpointer to the snetsta and stage tables
and the join key range 'time::endtime' matching
'ondate::offdate' between the stage dbptr and the
dlsensor tables. Provided the dlsensor table is
kept up-to-date we can use this table by itself
and compare directly with just the sitechan table.

Instead of generating and joining views, now just
generate a list of station dictionaries to allow a
simple 'if xxx in list' comparison. Much much faster
and simpler.

Also, account for new strong motion sensors (chans
HN.*) so modify the subset of channels to be a new
regex 'chan=~/^(BH|HN).*/'