Script to pull IRIS noise ranking of TA stations
Uses module BeautifulSoup. Scrapes all the data
files and html to create a JSON file of the
finalized data for all 450+ TA stations for
each of the channels BHE, BHN, and BHE.

At some point might like to move config to a
parameter file for portability. Leave for now
to concentrate on client-side interface
work.