Modify create_deploy_pointer
The deployment table should be considered the backbone
of deciding which stations are online or offline. Other
tables such as site or snetsta can have offdates/endtimes
at some arbitrary time in the future. This may mean an
invalid db table, but that is what they look like on our
systems. This commit is a work around that glitch. See
JIRA ticket ANZA-76 for more detail.